+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Primitives/KnownManagementTypes.cs
@@ -23,6 +23,8 @@ namespace Azure.Generator.Management.Primitives
     internal class KnownManagementTypes
     {
         public const string ArmResource = "Azure.ResourceManager.CommonTypes.Resource";
+        public const string ResourceUpdateModelId = "Azure.ResourceManager.Foundations.ResourceUpdateModel";
+
         public delegate MethodBodyStatement SerializationExpression(ValueExpression value, ScopedApi<Utf8JsonWriter> writer, ScopedApi<ModelReaderWriterOptions> options, SerializationFormat format);
         public delegate ValueExpression DeserializationExpression(CSharpType valueType, ScopedApi<JsonElement> element, SerializationFormat format);
