+++ eng/packages/http-client-csharp-mgmt/generator/Azure.Generator.Management/src/Visitors/NameVisitor.cs
@@ -5,8 +5,10 @@ using Azure.Core;
 using Azure.Generator.Management.Primitives;
 using Microsoft.TypeSpec.Generator.ClientModel;
 using Microsoft.TypeSpec.Generator.Input;
+using Microsoft.TypeSpec.Generator.Primitives;
 using Microsoft.TypeSpec.Generator.Providers;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
 namespace Azure.Generator.Management.Visitors;
@@ -15,11 +17,29 @@ internal class NameVisitor : ScmLibraryVisitor
 {
     private const string ResourceTypeName = "ResourceType";
 
+    private readonly HashSet<CSharpType> _resourceUpdateModelTypes = new HashSet<CSharpType>();
+
     protected override ModelProvider? PreVisitModel(InputModelType model, ModelProvider? type)
     {
         if (type is not null && TryTransformUrlToUri(model.Name, out var newName))
         {
             type.Update(name: newName);
+        }
+
+        if (type is not null && ManagementInputLibrary.IsResourceUpdateModel(model))
+        {
+            var enclosingResourceName = ManagementInputLibrary.FindEnclosingResourceNameForResourceUpdateModel(model);
+            var newModelName = $"{enclosingResourceName}Patch";
+
+            _resourceUpdateModelTypes.Add(type.Type);
+
+            type.Update(name: newModelName);
+
+            foreach (var serializationProvider in type.SerializationProviders)
+            {
+                serializationProvider.Update(name: newModelName);
+                _resourceUpdateModelTypes.Add(serializationProvider.Type);
+            }
         }
 
         // rename "Type" property to "ResourceType" in Azure.ResourceManager.CommonTypes.Resource
@@ -51,6 +71,25 @@ internal class NameVisitor : ScmLibraryVisitor
             }
         }
         return base.PreVisitModel(model, type);
+    }
+
+    protected override MethodProvider? VisitMethod(MethodProvider method)
+    {
+        var parameterUpdated = false;
+        foreach (var parameter in method.Signature.Parameters)
+        {
+            if (_resourceUpdateModelTypes.Contains(parameter.Type))
+            {
+                parameter.Update(name: "patch");
+                parameterUpdated = true;
+            }
+        }
+
+        if (parameterUpdated)
+        {
+            method.Update(signature: method.Signature);
+        }
+        return base.VisitMethod(method);
     }
 
     private bool TryTransformUrlToUri(string name, [MaybeNullWhen(false)] out string newName)
